// frontend/src/app/core/enums/ticket-status.enum.ts
export enum TicketStatus {
    PENDING = 'Pending',
    PENDING_MEMBERSHIP = 'Pending - Membership',
    SENT_TO_MUNICIPALITY = 'Sent to Municipality',
    CANNOT_TAKE = 'Cannot Take',
    NOT_ENGAGING = 'Not Engaging',
    DUPLICATE = 'Duplicate',
    CLOSED = 'Closed',
    COMPLETED = 'Completed',
    DISMISSED = 'Dismissed',
    REFUNDED = 'Refunded',
    GUILTY = 'Guilty',
    PAID_CREDIT_CARD = 'Paid - Credit Card',
    PAID_MEMBERSHIP = 'Paid - Membership',
    AWAITING_PAYMENT = 'Awaiting Payment',
}